requirements.txt
textblob
newspaper3k
gradio